% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dijkstra.R
\name{dijkstra}
\alias{dijkstra}
\title{Compute Dijkstra's algorithm}
\usage{
dijkstra(graph, init_node)
}
\arguments{
\item{graph}{A data.frame with three variables containing the edges of a graph and the weight of the edge.}

\item{init_node}{A node within \code{graph} from which to calculate the shortest path to every other node.}
}
\value{
A vector containing the shortest paths from \code{init_node} to every other node in the graph.
}
\description{
Compute Dijkstra's algorithm
}
\details{
This function computes Dijkstra's algorithm to find the shortest paths between nodes in a graph.
}
\examples{
dijkstra(wiki_graph, 1)
dijkstra(wiki_graph, 3) 
}
\references{
https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm (\href{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}{Wikipedia}).
}
