% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A weighted graph.}
\format{A data frame with 18 rows and 3 variables:
\describe{
 \item{v1}{nodes on one side of an edge}
 \item{v2}{nodes on the other side of the edge}
 \item{w}{the weight of the edge}
}}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset containing information on edges in a graph and their weight.
}
\keyword{datasets}
